[edit]プロジェクトの構成を修正し、不要なファイルを削除およびフォルダにまとめる

+++ src/Character/CharacterBase.cpp
@@ -212,7 +212,7 @@ bool CharacterBase::AttackHitGoodTiming(int attack_num)
 	// 終わってほしいフレーム
 	int end_time = m_attack_hit_damage[attack_num]->end_time;
 	// アニメーションの現在のフレーム
-	int play_anim_time = m_animation.m_contexts[0].play_time;
+	float play_anim_time = m_animation.m_contexts[0].play_time;
 	// starttime以上　かつ　endtime以下の時
 	if (start_time <= play_anim_time && play_anim_time <= end_time)
 	{
